First version to get to a prompt
Changes made so the program gets to the point where it can ask for
input from the player. There is only one room, so you can’t do much,
but this is a start.

Next step is to make the map and make it so the player can navigate it.